add name and role fields to user
This commit adds a full name and role (consumer or producer) to the user
model. It also adds forms for entering that information at signup, and
editing the name on the profile edit page. The role uses Rails 4.1's
enum construct, which you can access like so:

   user.consumer? # => true
   user.producer? # => false
   user.role # => "consumer"

The name is just a regular string.